Fixed MySQL configuration.
The MySQL configuration(which is never used), isn't configured
for the local instance of MySQL.  It needs to be configured so
that the database and user both exist.  I'm not sure why Laravel
needs a mysql configuration, but it's something that needs to be
explored in more detail later on.